pyproject.toml
[project]
name = "unitree-gallery-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
]
